Cluster nodes run as backup & prepared mission control for failover (#136)
- Added support for MissionParameters (parameters with values that can be changed at runtime) to MissionControl
- Updated DebugCanvas to display Role / ID / Render ID
- Centralized NodeRole into a new ClusterDisplay.Shared assembly so that both ClusterDisplay and Capcom can use it.
- Fixed bug in the UI causing it to sometime falsely report unaccounted for changes from the server in the launch configuration.
- L.aunchpad can now set its state to over instead of idle when aborting (to have a more similar behaviorbetween cases where a launchable finishes by itself vs is forcefully terminated).
- Centralized strings identifying collections in incrementalCollectionsUpdate.
- Centralized strings identifying objects in objectsUpdate.
- Moved Shared.LaunchCatalog.Constraints to Shared (since it is not also used for MissionParameters).
- Moved update version numbers out of MissionControlMirror to another private to Application object.

+++ source/com.unity.cluster-display/Runtime/MissionControl/Capsule/SendCapsuleStatus.cs
@@ -31,7 +31,7 @@ namespace Unity.ClusterDisplay.MissionControl.Capsule
 
         public async ValueTask Send(NetworkStream networkStream)
         {
-            CapsuleStatusMessage message = new() {NodeRole = (byte)NodeRole, NodeId = NodeId,
+            CapsuleStatusMessage message = new() {NodeRole = NodeRole, NodeId = NodeId,
                 RenderNodeId = RenderNodeId};
 
             // Send the message
